Implements user PG assignment and removal
Adds endpoints for assigning and removing users from a PG, including authorization checks.

Updates Spring Boot version and fixes minor code style issues.

The email field in the User entity is now unique.

+++ src/main/java/app/netlify/dsubha/service/UserPGService.java
@@ -12,6 +12,7 @@ import app.netlify.dsubha.entity.UserPG;
 
 @Component
 public class UserPGService {
+
 	@Autowired
 	UserRepository userRepository;
 
@@ -21,19 +22,36 @@ public class UserPGService {
 	@Autowired
 	UserPGRepository userPGRepository;
 
-	public UserPG assignUserToPg(String userId, String pgId) {
+	public UserPG assignUserToPg(String userId, String pgId, String adminId, boolean userJoined) {
 		User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
 
 		PG pg = pgRepository.findById(pgId).orElseThrow(() -> new RuntimeException("PG not found"));
 
-		UserPG userPG = new UserPG(user, pg);
+		if (!userJoined && !pg.getAdmin().get("userId").equals(adminId)) {
+			throw new RuntimeException("User is not authorized to add members");
+		}
+
+		UserPG userPG = new UserPG(user, pg, userJoined);
 		return userPGRepository.save(userPG);
 	}
 
-	public UserPG removeUserFromPG(String userId, String pgID) {
+	public UserPG removeUserFromPG(String userId, String pgID, String adminId) {
+		PG pg = pgRepository.findById(pgID).orElse(null);
 		UserPG userPG = userPGRepository.findByPgAndUser(pgID, userId);
 		if (userPG != null) {
+			if (!pg.getAdmin().get("userId").equals(adminId)) {
+				throw new RuntimeException("User is not authorized to remove members");
+			}
 			userPGRepository.delete(userPG);
+		}
+		return userPG;
+	}
+
+	public UserPG joinUserPG(String userId, String pgId) {
+		UserPG userPG = userPGRepository.findByPgAndUser(pgId, userId);
+		if (userPG != null) {
+			userPG.setUserJoined(true);
+			userPGRepository.save(userPG);
 		}
 		return userPG;
 	}
